kubectl: move events sorting interface to the api for general use
