+++ hack/.linted_packages
@@ -56,6 +56,7 @@ hack/cmd/teststale
 pkg/api
 pkg/api/annotations
 pkg/api/errors
+pkg/api/events
 pkg/api/install
 pkg/api/meta
 pkg/api/resource
