+++ pkg/api/events/sorted_event_list.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package kubectl
+package events
 
 import (
 	"k8s.io/kubernetes/pkg/api"
